Fix 10 WGSL shader compilation errors caught by gauntlet
Pre-existing WGSL bugs that hung the page on Dawn/D3D12 — none of these
were caused by the GLSL fix batches. Each effect now compiles cleanly.

Categorized:

1. `type` is a WGSL reserved keyword (3 effects)
   synth3d/flythrough3d, synth3d/fractal3d, mixer/patternMix
   Renamed the GPU uniform from `type` to `noiseType` (synth3d/*) and
   `patternType` (patternMix). The user-facing `type` global key in
   definition.js stays the same — DSL is unchanged. Only the GPU-side
   uniform identifier is renamed.

2. `i32 >> i32` requires u32 RHS in WGSL (3 effects, originally
   diagnosed as 2 — spookyTicker had it too)
   filter/glyphMap, filter/osd, filter/spookyTicker
   `row >> (n - x)` becomes `row >> u32(n - x)`. WGSL has no implicit
   integer-type coercion for shift operands.

3. WGSL forbids mixing `^` and `*` without parens (1 effect)
   filter/spookyTicker
   `u32(a) ^ u32(b) * c` becomes `u32(a) ^ (u32(b) * c)`. WGSL doesn't
   give XOR a precedence relative to multiplication.

4. `f32 + i32` mixed-type op (1 effect)
   render/pointsBillboardRender/deposit.wgsl
   `n + u.seed` (where seed is i32) becomes `n + f32(u.seed)`. Cast
   explicitly — WGSL has no implicit i32→f32 coercion.

5. textureSample inside non-uniform control flow (3 effects)
   filter/corrupt, filter/smooth, mixer/shadow
   These shaders sample inside if-blocks that depend on per-pixel
   values, which the WGSL spec forbids for plain textureSample (it
   needs uniform control flow for implicit mipmap derivative
   computation across a quad). Replaced with textureSampleLevel(...,
   0.0) which takes an explicit mip level and has no derivative
   requirement. None of these effects use mipmaps.

Verified before/after on Windows Chrome via the wgsl gauntlet:
all 10 previously-timing-out effects now compile in under 200ms each.

+++ shaders/effects/filter/glyphMap/wgsl/glyphMap.wgsl
@@ -149,7 +149,8 @@ fn glyphRow(g: i32, y: i32) -> i32 {
 // Return 1.0 if pixel (x, y) is set in glyph g, else 0.0
 fn glyphPixel(g: i32, x: i32, y: i32) -> f32 {
     let row = glyphRow(g, y);
-    let bit = (row >> (4 - x)) & 1;
+    // WGSL requires the right-hand side of `>>` to be a u32 (or vecN<u32>)
+    let bit = (row >> u32(4 - x)) & 1;
     return f32(bit);
 }
